backend/requirements_aggregator.txt
# Requirements for Data Aggregation and Database Population Script

# Database
psycopg2-binary==2.9.9

# Vector Database
pinecone-client==3.0.0

# AI/ML
openai==1.7.0

# Utilities
python-dotenv==1.0.0
